Broke out 'TagSim' class into 'TagSimClass.py'.
